axplat_aarch64_common

Module pl011

PL011 UART.

Functions

getchar

Reads a byte from the console, or returns None if no input is available.

init_early

Early stage initialization of the PL011 UART driver.

irq_handler

UART IRQ Handler

putchar

Writes a byte to the console.

read_bytes

Reads bytes from the console into the given mutable slice. Returns the number of bytes read.

write_bytes

Write a slice of bytes to the console.